+++ animation_nodes/data_structures/__init__.py
@@ -11,7 +11,7 @@ def importDataStructures():
     from . lists.base_lists import (
         Vector3DList, Vector2DList, Matrix4x4List, EdgeIndicesList, EulerList, ColorList,
         BooleanList, FloatList, DoubleList, LongList, IntegerList, UShortList, CharList,
-        QuaternionList, UIntegerList, ShortList, UShortList
+        QuaternionList, UIntegerList, ShortList, UShortList, Int2List
     )
 
     from . virtual_list.virtual_list import VirtualList, VirtualPyList