+++ animation_nodes/nodes/mesh/insert_custom_attribute.py
@@ -72,7 +72,7 @@ class InsertCustomAttributeNode(AnimationNode, bpy.types.Node):
             self.newInput(VectorizedSocket("Boolean", "useDataList",
             ("Value", "data"), ("Values", "data")))
         else:
-            self.newInput(VectorizedSocket("Edge Indices", "useDataList",
+            self.newInput(VectorizedSocket("Int2", "useDataList",
             ("Value", "data"), ("Values", "data")))